--- a/tests/Kernel/Types/string_test.cpp
+++ b/tests/Kernel/Types/string_test.cpp
@@ -64,6 +64,81 @@ TEST_CASE ("test append") {
   CHECK_EQ (str == string ("xyz"), true);
 }
 
+TEST_CASE ("test append") {
+  auto str= string ();
+  str << 'x';
+  CHECK (str == "x");
+  str << string ("yz");
+  CHECK (str == "xyz");
+}
+
+TEST_CASE ("test reserve along with append") {
+
+  SUBCASE ("reserved more space") {
+    auto str= string ();
+    str->reserve (6);
+    str << 'x';
+    CHECK_EQ (str == "x", true);
+    str << string ("yz");
+    CHECK_EQ (str == "xyz", true);
+    str << string (": larger than reserved space");
+    CHECK_EQ (str == "xyz: larger than reserved space", true);
+  }
+  SUBCASE ("reserved the same space") {
+    auto str= string ("abc");
+    str->reserve (3);
+    CHECK_EQ (str == "abc", true);
+  }
+  SUBCASE ("reserved less space should take no effect") {
+    auto str= string ("abc");
+    str->reserve (2);
+    CHECK_EQ (str == "abc", true);
+  }
+}
+
+TEST_CASE ("test expand_or_shrink_by along with sub index") {
+
+  SUBCASE ("expand") {
+    auto str       = string ("abc");
+    int  previous_n= str->expand_or_shrink_by (1);
+    CHECK_EQ (previous_n, 3);
+    str[3]= 'd';
+    CHECK_EQ (str == "abcd", true);
+  }
+  SUBCASE ("shrink") {
+    auto str       = string ("abc");
+    int  previous_n= str->expand_or_shrink_by (-1);
+    CHECK_EQ (previous_n, 3);
+    CHECK_EQ (str == "ab", true);
+  }
+  SUBCASE ("delta 0 takes no effect") {
+    auto str       = string ("abc");
+    int  previous_n= str->expand_or_shrink_by (0);
+    CHECK_EQ (previous_n, 3);
+    CHECK_EQ (str == "abc", true);
+  }
+}
+
+TEST_CASE ("test resize") {
+
+  SUBCASE ("expand") {
+    auto str= string ("abc");
+    str->resize (4);
+    str[3]= 'd';
+    CHECK_EQ (str == "abcd", true);
+  }
+  SUBCASE ("shrink") {
+    auto str= string ("abc");
+    str->resize (2);
+    CHECK_EQ (str == "ab", true);
+  }
+  SUBCASE ("delta 0 takes no effect") {
+    auto str= string ("abc");
+    str->resize (3);
+    CHECK_EQ (str == "abc", true);
+  }
+}
+
 /******************************************************************************
  * Conversions
  ******************************************************************************/