@da-liii 2024-07-13
在PR描述或者代码里面写一下，这个pr和https://github.com/XmacsLabs/lolly/pull/331有什么区别，如何做到避免内存问题的